+++ HISTORY.md
@@ -1,3 +1,8 @@
+0.0.4 / 2014-11-17
+==================
+
+* load ractive as npm module dependency
+* `pretty` and `ractiveFilename` options
 
 0.0.3 / 2014-11-16
 ==================